[BUGFIX] Set anchor in all preview links of content elements
The button »View webpage« in the list module opens a popup
with a link to the current page and an anchor to the
selected content element. This behaviour is expected as well
while selecting the »Show« option in the context menue of
content elements or the »Save and view page« button in
edit forms. Both places did not add the anchor.

Check the table and if it is »tt_content« then render the
same anchor in all three preview links.

Resolves: #81250
Releases: 7.6
Change-Id: I1ba806f37b01f79d7543444d22d157e8e5bc5f3a